# api/core.openapi.yaml
basePath: /
definitions:
...
  PrivilegeKey:
    enum:
    - core:clusters:read
    - core:clusters:write
    - core:allocations:read
    - core:allocations:write
    - core:projects:read
    - core:projects:write
    - core:users:read
    - core:users:write
    - core:organizations:read
    - core:organizations:write
    - core:traces:read
    - core:privileges:grant
    - core:roles:manage
    - core:events:manage
    type: string
    x-enum-varnames:
    - ClustersRead
    - ClustersWrite
    - AllocationsRead
    - AllocationsWrite
    - ProjectsRead
    - ProjectsWrite
    - UsersRead
    - UsersWrite
    - OrganizationsRead
    - OrganizationsWrite
    - TracesRead
    - PrivilegesGrant
    - RolesManage
    - EventsManage
...
  Role:
    properties:
      created_at:
        type: string
      description:
        type: string
        x-nullable: true
      id:
        type: string
      is_system:
        type: boolean
      name:
        type: string
    required:
    - created_at
    - id
    - is_system
    - name
    type: object
  RoleDetailResponse:
    properties:
      privileges:
        items:
          $ref: '#/definitions/PrivilegeKey'
        type: array
        x-nullable: true
      role:
        $ref: '#/definitions/Role'
    required:
    - role
    type: object
...
  UserRole:
    properties:
      granted_at:
        type: string
      granted_by:
        type: string
        x-nullable: true
      reason:
        type: string
        x-nullable: true
      role_id:
        type: string
      user_id:
        type: string
    required:
    - granted_at
    - role_id
    - user_id
    type: object
...
  createRoleRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
  grantRoleRequest:
    properties:
      reason:
        type: string
      role_id:
        type: string
    type: object
...
  revokeRoleRequest:
    properties:
      reason:
        type: string
    type: object
  rolePrivilegeRequest:
    properties:
      privilege:
        $ref: '#/definitions/PrivilegeKey'
    type: object
...
  updateRoleRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 'REST API for Apache Custos: organizations, users, projects, compute
    clusters, allocations, audit traces, and the privilege/role authorization layer.
    OIDC bearer auth required. Connector endpoints live under /connectors/<name>/.'
  title: Apache Custos Core API
  version: 0.2.0
paths:
...
  /roles:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Role'
            type: array
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Caller lacks roles:manage
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: List all roles
      tags:
      - Roles
    post:
      consumes:
      - application/json
      parameters:
      - description: Role payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/createRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Role'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "409":
          description: Role name collides
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a role
      tags:
      - Roles
  /roles/{id}:
    delete:
      description: System roles cannot be deleted. CASCADE drops every assignment
        of this role.
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: System role / unknown role
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a role
      tags:
      - Roles
    get:
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/RoleDetailResponse'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a role with its privilege bundle
      tags:
      - Roles
    put:
      consumes:
      - application/json
      description: System roles cannot be renamed.
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Role patch
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/updateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Role'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update role name / description
      tags:
      - Roles
  /roles/{id}/holders:
    get:
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/UserRole'
            type: array
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: List users holding the role
      tags:
      - Roles
  /roles/{id}/privileges:
    post:
      consumes:
      - application/json
      description: The new privilege propagates to every current holder.
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Privilege key
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rolePrivilegeRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Unknown key
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "409":
          description: Role already carries that privilege
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add a privilege to a role
      tags:
      - Roles
  /roles/{id}/privileges/{key}:
    delete:
      description: Removal propagates to every holder. Refuses to remove `privileges:grant`
        or `roles:manage` if that would leave no role anywhere carrying it.
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Privilege key
        enum:
        - core:clusters:read
        - core:clusters:write
        - core:allocations:read
        - core:allocations:write
        - core:projects:read
        - core:projects:write
        - core:users:read
        - core:users:write
        - core:organizations:read
        - core:organizations:write
        - core:traces:read
        - core:privileges:grant
        - core:roles:manage
        - core:events:manage
        in: path
        name: key
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Last-source guard or unknown key
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Role does not carry that privilege
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remove a privilege from a role
      tags:
      - Roles
...
  /users/{id}/roles:
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/UserRole'
            type: array
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: List roles a user holds
      tags:
      - Role Assignments
    post:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Role assignment payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/grantRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/UserRole'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "409":
          description: User already holds that role
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Grant a role to a user
      tags:
      - Role Assignments
  /users/{id}/roles/{roleId}:
    delete:
      consumes:
      - application/json
      description: Refuses if revoking would leave no holder of `privileges:grant`
        or `roles:manage` anywhere (last-meta-holder guard).
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: string
      - description: Optional reason
        in: body
        name: request
        schema:
          $ref: '#/definitions/revokeRoleRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Last-meta-holder guard
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: User does not hold that role
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Revoke a role from a user
      tags:
      - Role Assignments
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"